requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Deep Learning & RL
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# Data Sources
yfinance>=0.2.0

# Configuration
pyyaml>=6.0

# Hyperparameter Optimization
optuna>=3.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0

